+++ Alarm/AlarmDetailsViewController.swift
@@ -7,6 +7,9 @@
 
 import UIKit
 
+import FirebaseAuth
+import FirebaseDatabase
+
 class AlarmDetailsViewController: UIViewController {
 
     var alarm: Alarm!
@@ -18,7 +21,7 @@ class AlarmDetailsViewController: UIViewController {
         super.viewDidLoad()
 
         // Do any additional setup after loading the view.
-        self.alarmName.text = alarm.alarmName
+        self.alarmName.text = self.alarm.alarmName
         
         let utilities = Utilities()
         
@@ -39,5 +42,39 @@ class AlarmDetailsViewController: UIViewController {
         // Pass the selected object to the new view controller.
     }
     */
+    
+    @IBAction func deleteAlarm() {
+        
+        guard let userId = Auth.auth().currentUser?.uid else {
+            return
+        }
+        
+        // delete alarm in firebase
+        let ref = Database.database().reference()
+        
+        ref.child("/alarms/\(userId)/\(self.alarm.key!)").removeValue { error, ref in
+            if error == nil {
+                UNUserNotificationCenter.current().getPendingNotificationRequests { requests in
+                    for request in requests {
+                        
+                        // delete alarm in pending notifications (if it is enabled)
+                        if request.identifier == self.alarm.key! {
+                            if self.alarm.enabled == true {
+                                print("deleting data")
+                                UNUserNotificationCenter.current().removePendingNotificationRequests(withIdentifiers: [self.alarm.key!])
+                            }
+                            
+                            // if alarm is not enabled or disabled, it is not in the pending notification requests
+                            // hence there is no need to remove anything
+                        }
+                    }
+                    
+                    DispatchQueue.main.async {
+                        self.navigationController?.popToRootViewController(animated: true)
+                    }
+                }
+            }
+        }
+    }
 
 }